+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - changed: Simplify our React Native integration.
+- removed: Delete all rate plugins.
 
 ## 0.22.0 (2023-10-24)